chore: upgrade to nuxt3
